fix: fusion of init and join
