+++ src/commands/cli/templates/mod.rs
@@ -1,11 +1,7 @@
-mod default_global_config;
-mod default_local_config;
 mod general_config;
 mod pod_config;
 mod templates;
 
-pub use default_global_config::default_global_config;
-pub use default_local_config::default_local_config;
 pub use general_config::general_config;
 pub use pod_config::pod_config;
 pub use templates::templates;
