+++ src/commands/mod.rs
@@ -1,8 +1,13 @@
 pub mod cli;
 pub mod cli_commands;
+pub mod default_global_config;
+pub mod default_local_config;
 pub mod service;
 
 pub use cli_commands::{
-    Cli::{Init, Join, Start, Stop, Template},
+    Cli::{New, Start, Stop, Template},
     PodCommand,
 };
+
+pub use default_global_config::default_global_config;
+pub use default_local_config::default_local_config;
