+++ src/commands/service/mod.rs
@@ -1,11 +1,7 @@
-mod config;
-mod init;
-mod join;
+mod new;
 mod start;
 mod stop;
 
-pub use config::read_config;
-pub use init::init;
-pub use join::join;
+pub use new::new;
 pub use start::start;
 pub use stop::stop;
