+++ src/commands/service/new.rs
@@ -1,0 +1,81 @@
+use std::sync::Arc;
+
+use log::error;
+use tokio::{runtime::Runtime, sync::mpsc};
+
+use crate::{
+    commands::{
+        cli::cli_messager, cli_commands::PodArgs, default_global_config, default_local_config,
+        PodCommand,
+    },
+    config::{types::Config, GlobalConfig, LocalConfig},
+    network::server::Server,
+    pods::{declarations::Pod, whpath::WhPath},
+};
+
+pub async fn new(tx: mpsc::UnboundedSender<PodCommand>, args: PodArgs) -> Result<String, String> {
+    let (global_config, local_config, server, mount_point) = pod_value(&args).await;
+    let new_pod = match Pod::new(
+        args.name,
+        mount_point,
+        1,
+        global_config.general.peers,
+        server.clone(),
+        local_config.general.address,
+    )
+    .await
+    {
+        Ok(pod) => pod,
+        Err(e) => return Err(format!("Pod creation error: {}", e)),
+    };
+    match tx.send(PodCommand::JoinPod(new_pod)) {
+        Ok(_) => Ok("Pod created successfully".to_string()),
+        Err(e) => Err(format!("PodCommand send error: {}", e)),
+    }
+}
+
+fn add_hosts(mut global_config: GlobalConfig, additional_hosts: Vec<String>) -> GlobalConfig {
+    if additional_hosts.is_empty() {
+        return global_config;
+    }
+
+    if global_config.general.peers.is_empty() {
+        global_config.general.peers = additional_hosts;
+    } else {
+        for host in additional_hosts {
+            if !global_config.general.peers.contains(&host) {
+                global_config.general.peers.push(host);
+            }
+        }
+    }
+    global_config
+}
+
+async fn pod_value(args: &PodArgs) -> (GlobalConfig, LocalConfig, Arc<Server>, WhPath) {
+    let global_config: GlobalConfig =
+        GlobalConfig::read(".global_config.toml").unwrap_or(default_global_config());
+    let local_config: LocalConfig =
+        LocalConfig::read(".local_config.toml").unwrap_or(default_local_config(&args.name));
+    let server: Arc<Server> = Arc::new(Server::setup(&local_config.general.address).await);
+
+    let global_config = add_hosts(
+        global_config,
+        args.additional_hosts.clone().unwrap_or(vec![]),
+    );
+
+    let mount_point = args.path.clone().unwrap_or_else(|| ".".into());
+
+    (global_config, local_config, server, mount_point)
+}
+
+async fn pull_config(args: &PodArgs) -> Result<(), String> {
+    let rt = Runtime::new().unwrap();
+    // rt.block_on(cli_messager(
+    //     &args.url,
+    //     Cli::Init(PodArgs {
+    //         name: name,
+    //         path: Some(path.clone()),
+    //     }),
+    // ))?;
+    Ok(())
+}
